+++ src/integrationTest/java/eu/xenit/gradle/testrunner/AbstractIntegrationTest.java
@@ -31,19 +31,14 @@ public abstract class AbstractIntegrationTest {
     @Parameters(name = "Gradle v{0}")
     public static Collection<Object[]> testData() {
         return Arrays.asList(new Object[][]{
+                {"5.6"},
+                {"5.5.1"},
+                {"5.4.1"},
+                {"5.3.1"},
+                {"5.2.1"},
                 {"5.1.1"},
                 {"5.0"},
                 {"4.10.3"},
-                {"4.9"},
-                {"4.8.1"},
-                {"4.7"},
-                {"4.6"},
-                {"4.5.1"},
-                {"4.4.1"},
-                {"4.3.1"},
-                {"4.2"},
-                {"4.1"},
-                {"4.0.2"},
         });
     }