+++ src/main/java/eu/xenit/gradle/tasks/ResolveWarTask.java
@@ -1,11 +1,6 @@
 package eu.xenit.gradle.tasks;
 
-import org.apache.commons.io.FileUtils;
-import org.gradle.api.DefaultTask;
-import org.gradle.api.tasks.InputFile;
-import org.gradle.api.tasks.Internal;
-import org.gradle.api.tasks.OutputFile;
-import org.gradle.api.tasks.TaskAction;
+import static eu.xenit.gradle.alfresco.DockerAlfrescoPlugin.LABEL_PREFIX;
 
 import java.io.File;
 import java.io.IOException;
@@ -14,16 +9,23 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Supplier;
-
-import static eu.xenit.gradle.alfresco.DockerAlfrescoPlugin.LABEL_PREFIX;
+import org.apache.commons.io.FileUtils;
+import org.gradle.api.DefaultTask;
+import org.gradle.api.tasks.InputFile;
+import org.gradle.api.tasks.Internal;
+import org.gradle.api.tasks.OutputFile;
+import org.gradle.api.tasks.TaskAction;
 
 /**
  * Created by thijs on 31/05/17.
  * This task can get a configuration as input and resolves it to a file as output. It is mainly used to pass the
  * filename in the labels.
+ *
+ * @deprecated Since 3.x, will be removed in 5.0.0
  */
 @Deprecated
 public class ResolveWarTask extends DefaultTask implements WarEnrichmentTask {
+
     private Supplier<File> war;
     private List<Supplier<Map<String, String>>> labels = new ArrayList<>();