+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
  * [DEVEM-344](https://xenitsupport.jira.com/browse/DEVEM-344) - Make MergeWarTask extend the Zip task
     - Make it easier to publish WAR files with extensions applied
+ * [#44](https://github.com/xenit-eu/alfresco-docker-gradle-plugin/issues/44) - Automatically apply amps in the correct order
 
 ## Version 4.0.3 - 2019-01-17